+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (Unreleased)
 
 * Add `RSpec/RepeatedIncludeExample` cop. ([@biinari][])
+* Fix `RSpec/FilePath` when checking a file with a shared example. ([@pirj][])
 
 ## 1.43.1 (2020-08-17)